refactor(ws): split ledger boot-scan/GC internals into pane_ledger_scan submodule (1K-line file limit)
pane_ledger.rs had grown to 1060 lines, over the repo's 1K-lines-per-file
limit. Move the boot-scan/GC internals verbatim into a pane_ledger_scan
child module (the tabs_persist #[path] sibling-file convention):
BootScanReport, QuarantinedRow, and the boot_scan/gc/gc_locked/
quarantine_unparsable impl block. As a child module it shares the
parent's privacy scope, so the scan code keeps direct access to the
write-through index and private row-write helpers - no visibility
changes needed. Both types are re-exported from pane_ledger, so every
existing path compiles unchanged. No semantic changes.

pane_ledger.rs: 690 lines; pane_ledger_scan.rs: 385 lines.
